+++ app/build.gradle
@@ -5,7 +5,7 @@ android {
     buildToolsVersion '28.0.1'
 
     defaultConfig {
-        applicationId "com.google.android.gms.samples.vision.ocrreader"
+        applicationId "com.goverlabs.converter"
         minSdkVersion 14
         targetSdkVersion 26
         versionCode 1
