+++ app/src/main/java/activity/PermissionValidator.java
@@ -1,4 +1,4 @@
-package com.google.android.gms.samples.vision.ocrreader;
+package activity;
 
 import android.Manifest;
 import android.app.Activity;
@@ -6,6 +6,8 @@ import android.content.pm.PackageManager;
 import android.support.design.widget.Snackbar;
 import android.support.v4.app.ActivityCompat;
 import android.view.View;
+
+import com.goverlabs.converter.R;
 
 public class PermissionValidator {
